requirements.txt
Django~=3.0
DjangoRestFramework~=3.10
